feat(controllers): refactor import paths for error handling and authentication to use the new lib directory structure
