+++ src/components/auth/controller.ts
@@ -1,11 +1,11 @@
 import type { Request, Response } from "express";
 import * as service from "./service";
-import { normalizeError } from "@/utils/normalize-error";
+import { normalizeError } from "@/lib/normalize-error";
 import { sendErrorResponse } from "@/common/responses/error";
 import { sendSuccessResponse } from "@/common/responses/success";
 import { CustomError } from "@/common/custom/error";
 import { Role } from "@prisma/client";
-import { generateToken, comparePassword, hashPassword } from "@/utils/auth";
+import { generateToken, comparePassword, hashPassword } from "@/lib/auth";
 
 export const register = async (req: Request, res: Response) => {
   try {
