+++ src/components/properties/controller.ts
@@ -1,7 +1,7 @@
 import type { Request, Response } from "express";
 import prisma from "@/database/client";
 import * as service from "./service";
-import { normalizeError } from "@/utils/normalize-error";
+import { normalizeError } from "@/lib/normalize-error";
 import { sendErrorResponse } from "@/common/responses/error";
 import { sendSuccessResponse } from "@/common/responses/success";
 import { CustomError } from "@/common/custom/error";
