+++ src/components/tasks/controller.ts
@@ -2,8 +2,8 @@ import type { Request, Response } from "express";
 import * as service from "./service";
 import { sendErrorResponse } from "@/common/responses/error";
 import { sendSuccessResponse } from "@/common/responses/success";
-import { normalizeError } from "@/utils/normalize-error";
-import { parseAndValidateId } from "@/utils/utils";
+import { normalizeError } from "@/lib/normalize-error";
+import { parseAndValidateId } from "@/lib/utils";
 
 export const getTask = async (req: Request, res: Response) => {
   try {
